+++ Code/Libraries/Systems/Physics/MultiConvexMesh.cpp
@@ -160,19 +160,18 @@ Aabb SubConvexMesh::ComputeAabb(VertexArrayParam verts)
   if (!mValid)
     return mAabb;
 
-  for (uint i = 0; i < mIndices.Size(); ++i)
+  for (size_t i = 0; i < mIndices.Size(); ++i)
     mAabb.Expand(verts[mIndices[i]]);
   return mAabb;
 }
 
 void SubConvexMesh::Support(VertexArrayParam verts, Vec3Param direction, Vec3Ptr support)
 {
-  // Support should never be called on an invalid sub-mesh so no error checking
-  // should be needed
+  // Support should never be called on an invalid sub-mesh so no error checking should be needed
 
   // N version
   real longestDistance = -Math::PositiveMax();
-  for (uint i = 0; i < mIndices.Size(); ++i)
+  for (size_t i = 0; i < mIndices.Size(); ++i)
   {
     Vec3Param curr = verts[mIndices[i]];
     real dist = Math::Dot(direction, curr);
@@ -212,7 +211,7 @@ bool SubConvexMesh::CastRay(const Ray& localRay, MultiConvexMesh* mesh, ProxyRes
     {
       result.mTime = pointData.T;
       triangleWasHit = true;
-      result.ShapeIndex = i;
+      result.ShapeIndex = (uint)i;
       result.mPoints[0] = pointData.Points[0];
       result.mPoints[1] = pointData.Points[1];
       result.mContactNormal = tri.GetRawNormal();
@@ -221,7 +220,7 @@ bool SubConvexMesh::CastRay(const Ray& localRay, MultiConvexMesh* mesh, ProxyRes
   return triangleWasHit;
 }
 
-Triangle SubConvexMesh::GetTriangleIndexed(VertexArrayParam verts, uint index)
+Triangle SubConvexMesh::GetTriangleIndexed(VertexArrayParam verts, size_t index)
 {
   Vec3 p0 = verts[mTriangleIndices[index]];
   Vec3 p1 = verts[mTriangleIndices[index + 1]];
@@ -229,12 +228,12 @@ Triangle SubConvexMesh::GetTriangleIndexed(VertexArrayParam verts, uint index)
   return Triangle(p0, p1, p2);
 }
 
-Triangle SubConvexMesh::GetTriangle(VertexArrayParam verts, uint index)
+Triangle SubConvexMesh::GetTriangle(VertexArrayParam verts, size_t index)
 {
   return GetTriangleIndexed(verts, index * 3);
 }
 
-uint SubConvexMesh::GetTriangleCount()
+size_t SubConvexMesh::GetTriangleCount()
 {
   return mTriangleIndices.Size() / 3;
 }
@@ -266,8 +265,8 @@ void SubConvexMesh::DrawFaces(VertexArrayParam verts, Mat4Param transform, ByteC
   Debug::ActiveDebugConfig config;
   config.Alpha(100).Border(false).Filled(true);
 
-  uint triCount = GetTriangleCount();
-  for (uint i = 0; i < triCount; ++i)
+  size_t triCount = GetTriangleCount();
+  for (size_t i = 0; i < triCount; ++i)
   {
     Triangle tri = GetTriangle(verts, i);
     tri.p0 = Math::TransformPoint(transform, tri.p0);
@@ -282,8 +281,8 @@ void SubConvexMesh::DrawFaces(VertexArrayParam verts, Mat4Param transform, ByteC
 
 void SubConvexMesh::DrawEdges(VertexArrayParam verts, Mat4Param transform, ByteColor color)
 {
-  uint triCount = GetTriangleCount();
-  for (uint i = 0; i < triCount; ++i)
+  size_t triCount = GetTriangleCount();
+  for (size_t i = 0; i < triCount; ++i)
   {
     Triangle tri = GetTriangle(verts, i);
     tri.p0 = Math::TransformPoint(transform, tri.p0);
@@ -301,11 +300,11 @@ void SubConvexMesh::DrawEdges2d(VertexArrayParam verts, Mat4Param transform, Byt
   Debug::ActiveDebugConfig config;
   config.Alpha(100).Border(false).Filled(false);
 
-  uint faceIndexCount = mIndices.Size() / 2;
-  for (uint i = 0; i < faceIndexCount; ++i)
+  size_t faceIndexCount = mIndices.Size() / 2;
+  for (size_t i = 0; i < faceIndexCount; ++i)
   {
-    uint currIndex = i;
-    uint nextIndex = (i + 1) % faceIndexCount;
+    size_t currIndex = i;
+    size_t nextIndex = (i + 1) % faceIndexCount;
 
     Vec3 p0 = verts[mIndices[currIndex]];
     Vec3 p1 = verts[mIndices[nextIndex]];
@@ -389,8 +388,7 @@ void MultiConvexMesh::Serialize(Serializer& stream)
   SerializeNameDefault(mVertices, VertexArray());
   SerializeNameDefault(mMeshes, SubMeshArray());
 
-  // If we're loading then set the mesh owner on all of the sub-mesh bound
-  // arrays to ourself
+  // If we're loading then set the mesh owner on all of the sub-mesh bound arrays to ourself
   if (stream.GetMode() == SerializerMode::Loading)
   {
     for (size_t i = 0; i < mMeshes.Size(); ++i)
@@ -439,12 +437,11 @@ void MultiConvexMesh::CopyTo(MultiConvexMesh* destination)
     otherSubMesh->mIndices = mySubMesh->mIndices;
     otherSubMesh->mTriangleIndices = mySubMesh->mTriangleIndices;
   }
-  // There's no need to copy any other data (aabb, volume, etc...) because we
-  // mark the mesh as modified. When the next frame happens or the user tries to
-  // access any data the new mesh will be computed and all those values will be
-  // filled out. We could potentially save calculations by copying those values
-  // over, but if someone is cloning a mesh it's likely they will be modifying
-  // it shortly after anyways. Maybe optimize later...
+  // There's no need to copy any other data (aabb, volume, etc...) because we mark the mesh as
+  // modified. When the next frame happens or the user tries to access any data the new mesh will
+  // be computed and all those values will be filled out. We could potentially save calculations
+  // by copying those values over, but if someone is cloning a mesh it's likely they will be
+  // modifying it shortly after anyways. Maybe optimize later...
   destination->ResourceModified();
 }
 
@@ -452,12 +449,12 @@ void MultiConvexMesh::Draw(Mat4Param transform, bool drawEdges, bool drawFaces)
 {
   if (mFlags.IsSet(MultiConvexMeshFlags::EditType2D))
   {
-    for (uint i = 0; i < mMeshes.Size(); ++i)
+    for (size_t i = 0; i < mMeshes.Size(); ++i)
       mMeshes[i]->Draw2d(mVertices, transform, drawEdges, drawFaces);
   }
   else
   {
-    for (uint i = 0; i < mMeshes.Size(); ++i)
+    for (size_t i = 0; i < mMeshes.Size(); ++i)
       mMeshes[i]->Draw(mVertices, transform, drawEdges, drawFaces);
   }
 }
@@ -539,9 +536,8 @@ void MultiConvexMesh::UpdateAndNotifyIfModified()
   Validate(false);
   RebuildCachedInfo();
 
-  // As we have just changed values like center of mass and aabb, we need to
-  // notify any collider who used this mesh that we've changed so they can
-  // update world-space values.
+  // As we have just changed values like center of mass and aabb, we need to notify any collider
+  // who used this mesh that we've changed so they can update world-space values.
   ResourceEvent toSend;
   toSend.EventResource = this;
   DispatchEvent(Events::ResourceModified, &toSend);
@@ -578,8 +574,8 @@ Mat3 MultiConvexMesh::ComputeInvInertiaTensor(Vec3Param worldScale, real totalMa
 
   Mat3 totalInertiaTensor;
   totalInertiaTensor.ZeroOut();
-  // Combine all of the inertia tensors that were computed about their local
-  // center of masses to one inertia tensor about the total center of mass
+  // Combine all of the inertia tensors that were computed about their local center
+  // of masses to one inertia tensor about the total center of mass
   for (size_t i = 0; i < mMeshes.Size(); ++i)
   {
     SubConvexMesh* subMesh = mMeshes[i];
@@ -614,8 +610,7 @@ void MultiConvexMesh::RebuildCachedInfo()
 
 void MultiConvexMesh::ComputeCenterOfMassAndVolume()
 {
-  // If we're not valid then arbitrarily set the volume and center of mass to
-  // prevent crashes.
+  // If we're not valid then arbitrarily set the volume and center of mass to prevent crashes.
   if (!GetValid())
   {
     mVolume = 1;
@@ -626,7 +621,7 @@ void MultiConvexMesh::ComputeCenterOfMassAndVolume()
   mCenterOfMass = Vec3::cZero;
   mVolume = real(0.0);
   // Accumulate the volume and weighted center of mass of each sub shape
-  for (uint i = 0; i < mMeshes.Size(); ++i)
+  for (size_t i = 0; i < mMeshes.Size(); ++i)
   {
     SubConvexMesh* subMesh = mMeshes[i];
     subMesh->ComputeCenterOfMassAndVolume(mVertices);
@@ -645,7 +640,7 @@ Aabb MultiConvexMesh::ComputeAabb()
 {
   mAabb.SetInvalid();
 
-  for (uint i = 0; i < mMeshes.Size(); ++i)
+  for (size_t i = 0; i < mMeshes.Size(); ++i)
     mAabb.Combine(mMeshes[i]->ComputeAabb(mVertices));
   return mAabb;
 }
