+++ Code/Libraries/Systems/Physics/PhysicsRaycastProvider.cpp
@@ -53,7 +53,7 @@ void PhysicsRaycastProvider::RayCast(Ray& ray, CastInfo& castInfo, RaycastResult
   filter.SetState(BaseCastFilterFlags::IgnoreStatic, !mStaticColliders);
 
   // Form a physics cast result with the total result's capacity
-  CastResults physicsResults(results.mCapacity, filter);
+  CastResults physicsResults((uint)results.mCapacity, filter);
 
   // Cast the ray into physics
   physicsSpace->CastRay(ray, physicsResults);
@@ -75,8 +75,7 @@ void PhysicsRaycastProvider::FrustumCast(Frustum& frustum, CastInfo& castInfo, R
 
   CastFilter filter;
 
-  // Set up the filters for multi casting (make sure to get the flags that still
-  // matter such as ghost)
+  // Set up the filters for multi casting (make sure to get the flags that still matter such as ghost)
   if (!mSelectGhosts)
     filter.Set(BaseCastFilterFlags::IgnoreGhost);
   if (!mDynamicColliders)
@@ -87,7 +86,7 @@ void PhysicsRaycastProvider::FrustumCast(Frustum& frustum, CastInfo& castInfo, R
     filter.Set(BaseCastFilterFlags::IgnoreKinematic);
 
   // Form a physics cast result with the total result's capacity
-  CastResults physicsResults(results.mCapacity, filter);
+  CastResults physicsResults((uint)results.mCapacity, filter);
   // Cast to get the results from physics
   physicsSpace->CastFrustum(frustum, physicsResults);
   // And then add all of the results that we got
